Includes `download_type` parameter to select the type of download job
