+++ mash/services/jobcreator/base_job.py
@@ -69,6 +69,7 @@ class BaseJob(object):
         self.image_description = kwargs.get('image_description')
         self.use_build_time = kwargs.get('use_build_time')
         self.force_replace_image = kwargs.get('force_replace_image')
+        self.download_type = kwargs.get('download_type', 'OBS')
         self.kwargs = kwargs
 
         if self.raw_image_upload_type and self.last_service == 'upload':
@@ -126,10 +127,16 @@ class BaseJob(object):
                 self.conditions_wait_time
 
         if self.disallow_licenses:
-            download_message['download_job']['disallow_licenses'] = self.disallow_licenses
+            download_message['download_job']['disallow_licenses'] = \
+                self.disallow_licenses
 
         if self.disallow_packages:
-            download_message['download_job']['disallow_packages'] = self.disallow_packages
+            download_message['download_job']['disallow_packages'] = \
+                self.disallow_packages
+
+        if self.download_type:
+            download_message['download_job']['download_type'] = \
+                self.download_type
 
         return JsonFormat.json_message(download_message)
